refactor(core/categories): relocate validateNewCategoryName + RESERVED_TOKENS to core; update architecture.md (Story B)
- New src/core/categories/category-name.ts: pure domain validator, no external deps
- src/cli/utils/interactive.ts: import + re-export from @core/categories/category-name.js
- docs/architecture.md: add core/categories/ entry between ingest/ and ledger/

No behaviour change; all 461 tests stay green (R10).
